requirements.txt
fastapi==0.115.5
python-multipart==0.0.12
uvicorn[standard]==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
trafilatura>=1.8.0
openai==1.54.3
chromadb==0.5.15
